requirements.txt
aiofiles==25.1.0
anyio==4.11.0
asyncpg==0.30.0
bcrypt==4.1.2
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
coverage==7.12.0
deptry==0.24.0
Flask==3.1.2
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.3.1
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
priority==2.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.8.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-dotenv==1.2.1
Quart==0.20.0
quart-cors==0.8.0
requests==2.32.5
requests-oauthlib==2.0.0
requirements-parser==0.13.0
rsa==4.9.1
sniffio==1.3.1
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.3.1
